Change radioBroadcastPower_dbm from uint16_t to uint8_t

+++ Firmware/LoRaSerial_Firmware/settings.h
@@ -131,7 +131,7 @@ typedef struct struct_settings {
   bool encryptData = true; //AES encrypt each packet
   uint8_t encryptionKey[16] = { 0x37, 0x78, 0x21, 0x41, 0xA6, 0x65, 0x73, 0x4E, 0x44, 0x75, 0x67, 0x2A, 0xE6, 0x30, 0x83, 0x08 };
   bool dataScrambling = false; //Use IBM Data Whitening to reduce DC bias
-  uint16_t radioBroadcastPower_dbm = 30; //Transmit power in dBm. Max is 30dBm (1W), min is 14dBm (25mW).
+  uint8_t radioBroadcastPower_dbm = 30; //Transmit power in dBm. Max is 30dBm (1W), min is 14dBm (25mW).
   float frequencyMin = 902.0; //MHz
   float frequencyMax = 928.0; //MHz
   uint8_t numberOfChannels = 50; //Divide the min/max freq band into this number of channels and hop between.
